refactor(evaluation): extract resolve_project_name into helpers module
Moves the previously-private `_resolve_project_name` out of evaluator.py
into `opik.evaluation.helpers.resolve_project_name`, and moves its unit
tests into a dedicated `test_helpers.py`. Tests exercise the helper with
raw values, no mocks required.
