+++ sdks/python/tests/unit/evaluation/test_evaluate.py
@@ -3622,62 +3622,6 @@ class TestMergeBlueprintIntoConfig:
         assert result["agent_configuration"] == {"_blueprint_id": "bp-456"}
 
 
-class TestResolveProjectName:
-    def test_dataset_has_no_project__user_value_used(self, capture_log):
-        resolved = evaluator_module._resolve_project_name(
-            value_from_dataset=None,
-            value_from_user="caller-project",
-            caller_name="evaluate",
-        )
-
-        assert resolved == "caller-project"
-        assert capture_log.records == []
-
-    def test_dataset_has_no_project__user_none__returns_none(self, capture_log):
-        resolved = evaluator_module._resolve_project_name(
-            value_from_dataset=None,
-            value_from_user=None,
-            caller_name="evaluate",
-        )
-
-        assert resolved is None
-        assert capture_log.records == []
-
-    def test_dataset_has_project__user_none__returns_dataset_project__no_warning(
-        self, capture_log
-    ):
-        resolved = evaluator_module._resolve_project_name(
-            value_from_dataset="dataset-project",
-            value_from_user=None,
-            caller_name="evaluate",
-        )
-
-        assert resolved == "dataset-project"
-        assert capture_log.records == []
-
-    def test_dataset_has_project__user_override__dataset_wins_and_warning_logged(
-        self, capture_log
-    ):
-        resolved = evaluator_module._resolve_project_name(
-            value_from_dataset="dataset-project",
-            value_from_user="caller-project",
-            caller_name="evaluate_prompt",
-        )
-
-        assert resolved == "dataset-project"
-        warning_records = [
-            record
-            for record in capture_log.records
-            if record.levelno == logging.WARNING
-        ]
-        assert len(warning_records) == 1
-        message = warning_records[0].getMessage()
-        assert "deprecated" in message
-        assert "evaluate_prompt()" in message
-        assert "dataset-project" in message
-        assert "caller-project" in message
-
-
 def test_evaluate__dataset_has_project_name__caller_override_ignored_and_warning_logged(
     fake_backend, capture_log
 ):
